Refactor : pages의 edit 폴더 project로 통합
