+++ src/components/detail/index.ts
@@ -1,1 +1,2 @@
 export { default as DetailDiv } from './DetailDiv';
+export { default as DetailEdit } from './DetailEdit';
