--- src/pages/edit/index.ts
@@ -1,2 +1,0 @@
-export { default as DetailEdit } from './DetailEdit';
-export { default as ProjectAddPage } from './ProjectAddPage';
