+++ src/pages/index.tsx
@@ -3,8 +3,7 @@ import 정리용 페이지
 import {HomePage, NavPage, PageNotFound} from './components'    
 */
 export { HomePage } from './home';
-export { NavPage, ProjectPage, ProjectDetailPage, AboutMePage } from './nav';
-export { DetailEdit, ProjectAddPage } from './edit';
+export { NavPage, ProjectPage, ProjectDetailPage, AboutMePage, ProjectAddPage } from './nav';
 export { PageNotFound } from './PageNotFound';
 export { Login, SignUp, Spinner } from './login';
 export { ContactMe } from './contact/ContactMe';
