+++ src/pages/nav/index.tsx
@@ -2,3 +2,4 @@ export { default as NavPage } from './NavPage';
 export { default as ProjectDetailPage } from './ProjectDetailPage';
 export { default as ProjectPage } from './ProjectPage';
 export { default as AboutMePage } from './AboutMePage';
+export { default as ProjectAddPage } from './ProjectAddPage';
